feat: NavBar-flexible-for-any-links
In App.tsx:

Created an array of links with objects containing the text and onClick properties for each navigation link.
Passed the links array as a prop to the Navbar component.

In Navbar.tsx:

Updated the interface NavbarProps to expect an array of objects with text and onClick properties.
Mapped over the links array using object destructuring to access the text and onClick properties correctly.
Passed the onClick function to the NavLink component as a prop.

In NavLink.tsx:

Updated the interface NavLinkProps to expect an onClick function.
Destructured the onClick and children properties from the props.
Passed the onClick function to the onClick event handler of the anchor element.